Split UserController and UserRouter
